Improve nix-shell environment
this is an improvement, but once in a nix shell cabal still wants to build
every source-repository-package block despite the pacakge db containing
all the packages neccessary.

Verified that plutus and plutus-stater are using the same cabal-install.
The hack to get around this is that once the user has entered a
nix-shell one can comment out all the source-repository-package sections
in cabal.project and proceed to build the project regularly with cabal
and it all builds fine.

TODO: Figure out if cabal has a way to search the package-db first
before diving into source-repository-package sections